lib: fix comparison against unsigneds (related to mime_get_length)
CID 1442412
CID 1442448
CID 1442707
CID 1443808
CID 1443828
CID 1443863
CID 1443866
CID 1443873
CID 1443875

+++ lib/email/mime.cpp
@@ -2094,7 +2094,7 @@ BOOL mime_check_dot(MIME *pmime)
  *	@return
  *		length of mime object
  */
-long mime_get_length(MIME *pmime)
+ssize_t mime_get_length(MIME *pmime)
 {
 	int		tag_len, val_len;
 	size_t	mime_len, tmp_len;
